requirements.txt
matplotlib==3.7.2
pytz==2023.3
scikit-learn==1.3.0
torch==2.0.1
yfinance==0.2.28

torchvision==0.15.2
Pillow==9.5.0
